+++ NEWS.md
@@ -1,3 +1,7 @@
+# DrBristol (Version 0.0.1.3000)
+
+* Handles multiple kinds of evidence against multiple rivals including overlapping pieces of evidence.
+
 # DrBristol (Version: 0.0.1.1000)
 
 * Adding ability to specify size of rival urn to enable omnibus tests of working versus any rival including sensitivity analysis.